reformat fee payment on sync
