+++ app/codes/state_updater.py
@@ -8,7 +8,7 @@ import traceback
 from app.codes.helpers.FetchRespository import FetchRepository
 from app.codes.helpers.TransactionCreator import TransactionCreator
 
-from app.codes.transactionmanager import Transactionmanager
+from .transactionmanager import Transactionmanager, get_valid_addresses
 from .helpers.SmartContractStateValidator import validate
 from app.codes.clock.global_time import get_corrected_time_ms
 from app.codes.helpers.CentralRespository import CentralRepository
@@ -16,9 +16,10 @@ from .db_updater import *
 from app.codes.networkscoremanager import get_invalid_block_creation_score, get_invalid_receipt_score, get_valid_block_creation_score, get_valid_receipt_score, update_network_trust_score_from_receipt
 from .p2p.utils import get_peers
 
-from app.nvalues import NETWORK_TRUST_MANAGER_PID, set_attr
+from ..nvalues import NETWORK_TRUST_MANAGER_PID, TREASURY_WALLET_ADDRESS
 
-from ..constants import COMMITTEE_SIZE, INITIAL_NETWORK_TRUST_SCORE, NEWRL_DB
+
+from ..constants import ALLOWED_FEE_PAYMENT_TOKENS, COMMITTEE_SIZE, INITIAL_NETWORK_TRUST_SCORE, NEWRL_DB
 from ..ntypes import BLOCK_VOTE_MINER, NEWRL_TOKEN_CODE, NEWRL_TOKEN_NAME, TRANSACTION_MINER_ADDITION, TRANSACTION_ONE_WAY_TRANSFER, TRANSACTION_SC_UPDATE, TRANSACTION_SMART_CONTRACT, TRANSACTION_TOKEN_CREATION, TRANSACTION_TRUST_SCORE_CHANGE, TRANSACTION_TWO_WAY_TRANSFER, TRANSACTION_WALLET_CREATION
 
 logger = logging.getLogger(__name__)
@@ -58,14 +59,17 @@ def update_db_states(cur, block):
             'trans_code']
 
         try:
-            update_state_from_transaction(
-                cur,
-                transaction['type'],
-                transaction_data,
-                transaction_code,
-                transaction['timestamp'],
-                signature
-            )
+            if pay_fee_for_transaction(cur, transaction):
+                update_state_from_transaction(
+                    cur,
+                    transaction['type'],
+                    transaction_data,
+                    transaction_code,
+                    transaction['timestamp'],
+                    signature
+                )
+            else:
+                logger.info(f'Fee payment failed for transaction {transaction_code}')
         except Exception as e:
             logger.error(f'Error in transaction: {str(transaction)}')
             logger.error(str(e))
@@ -308,3 +312,37 @@ def get_value_txns(transaction_signer, transaction_data):
         value_txns_local.append(transfer_proposal.get_transaction_complete())
 
     return value_txns_local
+
+
+def get_fees_for_transaction(transaction):
+    if 'fee' in transaction:
+        return transaction['fee']
+    else:
+        return 0
+
+
+def pay_fee_for_transaction(cur, transaction):
+    fee = get_fees_for_transaction(transaction)
+
+    # Check for 0 fee transactions and deprioritize accordingly
+    if fee == 0:
+        return True
+
+    currency = transaction['currency']
+    if currency not in ALLOWED_FEE_PAYMENT_TOKENS:
+        return False
+
+    payers = get_valid_addresses(transaction)
+
+    for payee in payers:
+        balance = get_wallet_token_balance(cur, payee, currency)
+        if balance < fee / len(payers):
+            return False
+        transfer_tokens_and_update_balances(
+            cur,
+            payee,
+            TREASURY_WALLET_ADDRESS,
+            transaction['currency'],
+            fee / len(payers)
+        )
+    return True
